Объединены компоненты в один
