--- packages/client/src/components/PageError/pageError.module.scss
@@ -1,27 +1,0 @@
-.container {
-  display: flex;
-  align-items: center;
-  flex-direction: row;
-  color: #fff;
-}
-.code {
-  font-size: 36px;
-  line-height: 48px;
-}
-
-.divider {
-  background: #fff;
-  width: 1px;
-  margin: 0 25px;
-  height: 50px;
-}
-
-.message {
-  font-size: 24px;
-  line-height: 34px;
-}
-
-.discription {
-  font-size: 16px;
-  color: #ACACAC
-}
