+++ packages/client/src/pages/Error/errorPage.module.scss
@@ -1,7 +1,35 @@
-.container {
+.pageContainer {
   height: 100vh;
   display: flex;
   align-items: center;
   justify-content: center;
-  background: linear-gradient(180deg, #3B5E53 0%, #091E20 100%);
+  background: linear-gradient(180deg, #3b5e53 0%, #091e20 100%);
+}
+
+.container {
+  display: flex;
+  align-items: center;
+  flex-direction: row;
+  color: #fff;
+}
+.code {
+  font-size: 36px;
+  line-height: 48px;
+}
+
+.divider {
+  background: #fff;
+  width: 1px;
+  margin: 0 25px;
+  height: 50px;
+}
+
+.message {
+  font-size: 24px;
+  line-height: 34px;
+}
+
+.discription {
+  font-size: 16px;
+  color: #acacac;
 }
